% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{cl_mututal_exclusivity}
\alias{cl_mututal_exclusivity}
\title{Calculate mutual exclusivity}
\usage{
cl_mututal_exclusivity(
  se,
  mutex_output_exists,
  mutex_path = "/",
  min_variants = 10,
  prior_modules = NULL,
  save = "/home/metadata"
)
}
\arguments{
\item{se}{SE object}

\item{mutex_output_exists}{logical specifying if mutex output already exists}

\item{mutex_path}{path for mutex software}

\item{min_variants}{minimum number of patients mutated for analysing for
mutual exclusivity}

\item{prior_modules}{character vector of prior module to add to the modules}

\item{save}{path for existing mutex output}
}
\value{
SE object with metadata containing subtype enrichment tests
'me_modules' and 'ngs_exclusive'
}
\description{
Run the mutex algorithm on the mutational data
}
\examples{
\dontrun{
  sim_data <- sim_data()

  se <- prepare_data(
    data = sim_data,
    vars = c("OS", "DFS"),
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "PatientID",
    treatment_column = "Adj",
    mutation_columns = colnames(sim_data)[14:ncol(sim_data)]
  )


  se <- cl_mututal_exclusivity(se,
    min_variants = 10,
    mutex_output_exists = TRUE,
    save = "metadata/",
    mutex_path = "path/to/mutex/installation"
  )
}

}
