% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{cl_subtype_enrichment}
\alias{cl_subtype_enrichment}
\title{Calculate subtype enrichment}
\usage{
cl_subtype_enrichment(se, col_label, row_label, digits)
}
\arguments{
\item{se}{SE object}

\item{col_label}{name of subtype 1}

\item{row_label}{name of subtype 2}

\item{digits}{how many digits for the reported p-value}
}
\value{
SE object with metadata containing subtype enrichment
tests 'subtype_enrichment'
}
\description{
Calculating hypergeometric tests for two subtype columns to check for
associations.
}
\examples{
{

simed_data <- sim_data()
  
se <- prepare_data(
    data = simed_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(simed_data)[1:5])
    
se <- cl_subtype_enrichment(
    se = se,
    col_label = "s",
    row_label = "r",
    digits = 3)
  
}

}
