% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{cl_treatment_specific_biomarkers}
\alias{cl_treatment_specific_biomarkers}
\title{Calculate treatment-specific biomarkers}
\usage{
cl_treatment_specific_biomarkers(
  se,
  include_covariates = NULL,
  min_samples = 10,
  min_redistribution = 3,
  treatment = NULL,
  subtypes = NULL,
  readouts = c("OS")
)
}
\arguments{
\item{se}{SE object}

\item{include_covariates}{character vector with confounding factors}

\item{min_samples}{minimum number of mutants for a mutation to be
included in statistical tests}

\item{min_redistribution}{minimum number of redistributed tumours per
gene module for being included}

\item{treatment}{character for treatment arm column}

\item{subtypes}{character for subtype column}

\item{readouts}{character for readout (OS,PFS,DFS,RFS,ORR)}
}
\value{
SE object with metadata containing treatment-specific statistical
tests 'treatment_specific_biomarkers'
}
\description{
Calculate biomarkers for each readout, subtype and treatment arm
}
\examples{
{
  sim_data <- sim_data()
  
  se <- prepare_data(
    data = sim_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(sim_data)[1:5])
    
    
  se <- cl_treatment_specific_biomarkers(se, 
     min_samples = 10,
     min_redistribution = 3,
     treatment = c("chemo", "no_chemo"),
     subtypes = c("s", "r"),
     readouts = c("OS"))
  
}

}
