% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{cl_tsb_modules_oncoprint}
\alias{cl_tsb_modules_oncoprint}
\title{Calculate module visualisations for treatment-specific biomarkers}
\usage{
cl_tsb_modules_oncoprint(se, p_value, fdr)
}
\arguments{
\item{se}{SE object}

\item{p_value}{p-value threshold}

\item{fdr}{FDR threshold}
}
\value{
list of ggplot objects
}
\description{
Calculate forest plots, heatmaps and oncoprints for
treatment-specific biomarkers
}
\examples{
{

  sim_data <- sim_data()
  
  se <- prepare_data(
    data = sim_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(sim_data)[1:5])
    
    
  se <- cl_treatment_specific_biomarkers(se, 
     min_samples = 10,
     min_redistribution = 3,
     treatment = c("chemo", "no_chemo"),
     subtypes = c("s", "r"),
     readouts = c("OS"))
  
  se <- pl_treatment_specific_biomarkers_subtype(se, fdr_max = 0.1)
  
  se <- cl_tsb_modules_oncoprint(se, p_value = 0.05, fdr = 0.1)
  
  metadata(se)$tsb_modules_oncoprint$OS$`chemo/resistant`
}

}
