% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{make_table_summary}
\alias{make_table_summary}
\title{Make summary table}
\usage{
make_table_summary(
  cond = conditions,
  preds = predictive_biomarkers_forest,
  tsbiomarkers = treatment_specific_biomarkers_forest,
  clin = data_clinical,
  mut = data_mutations,
  mut_me = data_mutations_me,
  fdr_int_threshold = 0.1,
  fdr_biomarker_threshold = 0.1,
  round_digit = 2,
  eps = 1e-05,
  sample_column = "sample",
  covariates = NULL,
  treatment = NULL
)
}
\arguments{
\item{cond}{specify all readouts and subtypes}

\item{preds}{association tests for interactions after getting effect sizes
and confidence intervals}

\item{tsbiomarkers}{association tests for treatment-specific biomarkers
after getting effect sizes and confidence intervals}

\item{clin}{clinical dataframe}

\item{mut}{mutational dataframe}

\item{mut_me}{mutational dataframe containing mutually exclusive modules}

\item{fdr_int_threshold}{interaction FDR threshold}

\item{fdr_biomarker_threshold}{teatment-specific FDR threshold}

\item{round_digit}{rounding digit}

\item{eps}{numerical tolerance of numerical values}

\item{sample_column}{column in clinical and mutational data specifying the
individual}

\item{covariates}{confounding factors}

\item{treatment}{treatment column}
}
\value{
summary table dataframe
}
\description{
make the summary table from the previous tests
}
\examples{
{

  sim_data <- sim_data()
  
  se <- prepare_data(
    data = sim_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(sim_data)[1:5])
    
    
  se <- cl_treatment_specific_biomarkers(se, 
     min_samples = 10,
     min_redistribution = 3,
     treatment = c("chemo", "no_chemo"),
     subtypes = c("s", "r"),
     readouts = c("OS"))
  
  se <- pl_treatment_specific_biomarkers_subtype(se, fdr_max = 0.1)
  
  se <- cl_tsb_modules_oncoprint(se, p_value = 0.05, fdr = 0.1)
  
  se <- cl_predictive_biomarkers(
    se = se,
    min_samples = 10,
    min_redistribution = 5)
  
  se <- pl_predictive_biomarkers_subtypes(se,
    fdr_max = 0.1,
    colors = NULL,
    labels = NULL)
  
  se <- cl_predictive_comparison(
    se = se,
    subtypes = c("s", "r"),
    readouts = c("OS"),
    covariates = FALSE,
    fdr_i = 0.5)
    
  summary_table <- make_table_summary(
    cond = metadata(se)$conditions,
    preds = metadata(se)$predictive_biomarkers_forest,
    tsbiomarkers = metadata(se)$treatment_specific_biomarkers_forest,
    clin = rowData(se),
    mut = assay(se),
    mut_me = metadata(se)$me_modules,
    fdr_int_threshold = 0.4,
    fdr_biomarker_threshold = 0.4,
    round_digit = 2,
    sample_column = "sample",
    covariates = c(1),
    treatment = metadata(se)$wf_meta$treatment
)
}

}
