% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoBird.R
\docType{package}
\name{OncoBird}
\alias{OncoBird}
\title{The OncoBird package}
\description{
\code{OncoBird} identifies candidates for predictive biomarkers in oncology
trials by leveraging mutually exclusive somatic mutations in tumour subtypes
for subgroup analysis. It systematically screens for predictive effects of
somatic alterations in pre-defined tumour subtypes. It allows users to gain
insights about the biomarker landscape in their trial and explore putative
cancer subtypes and how they may be leveraged for precision oncology.
In our manuscript, we showcased \code{OncoBird} for a clinical trial for
FOLFIRI plus either cetuximab and bevacizumab in metastatic colorectal
cancer. Here, we showcase \code{OncoBird} by using a publicly available
clinical trial for gefitinib in non-small cell lung cancer.
}
\author{
Alexander J Ohnmacht \email{ohnmachtalexander@gmail.com} and
Michael P Menden \email{michael.menden@helmholtz-muenchen.de}
}
