% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{OncoBird_example}
\alias{OncoBird_example}
\title{Get small test dataset for OncoBird}
\usage{
OncoBird_example(path = NULL)
}
\arguments{
\item{path}{Path to the dummy dataset}
}
\value{
If path = NULL then it returns all datasets in extdata,
if path = filename then it returns paths to file
}
\description{
Derive treatment effect adjusted confidence intervals
}
\examples{
{

  OncoBird_example()
  sim_data <- sim_data()
  str(sim_data)
  
}
}
