% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{pl_enrichment_genomics}
\alias{pl_enrichment_genomics}
\title{Plot enrichment for mutations in subtypes}
\usage{
pl_enrichment_genomics(se, p_value = 0.05, fdr = 0.2)
}
\arguments{
\item{se}{SE object}

\item{p_value}{p-value threshold for plotting results}

\item{fdr}{false discovery rate threshold for reporting results}
}
\value{
ggplot object
}
\description{
Plotting results after 'cl_enrichment_genomics'
}
\examples{
{

  sim_data <- sim_data()
  
  se <- prepare_data(
    data = sim_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(sim_data)[1:5])
    
  se <- cl_subtype_enrichment(
    se = se,
    col_label = "s",
    row_label = "r",
    digits = 3
  )
  
  se <- cl_enrichment_genomics(se,
    sample_column = "sample",
    min_mutants = 10,
    subtype = c("s", "r")
  )
  
  # no siginficants expected
  pl_enrichment_genomics(se, p_value = 0.5, fdr = 1)
  
}

}
