% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{pl_example}
\alias{pl_example}
\title{Plot examples}
\usage{
pl_example(
  se,
  mutations,
  readout,
  subtype_column,
  subtype,
  treatment = NULL,
  covariates = c(1)
)
}
\arguments{
\item{se}{SE object}

\item{mutations}{character vector of mutations}

\item{readout}{character readout}

\item{subtype_column}{character subtype name}

\item{subtype}{character subtype level name}

\item{treatment}{character treatment level}

\item{covariates}{character vector of confounding factors}
}
\value{
list of ggplot objects
}
\description{
Visualise a specific example for the presented analysis steps
}
\examples{
{

  sim_data <- sim_data()
  
  se <- prepare_data(
    data = sim_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(sim_data)[1:5])
    
    
  se <- cl_treatment_specific_biomarkers(se, 
     min_samples = 10,
     min_redistribution = 3,
     treatment = c("chemo", "no_chemo"),
     subtypes = c("s", "r"),
     readouts = c("OS"))
  
  se <- pl_treatment_specific_biomarkers_subtype(se, fdr_max = 0.1)
  
  se <- cl_tsb_modules_oncoprint(se, p_value = 0.05, fdr = 0.1)
  
  se <- cl_predictive_biomarkers(
    se = se,
    min_samples = 10,
    min_redistribution = 5)
  
  se <- pl_predictive_biomarkers_subtypes(se,
    fdr_max = 0.1,
    colors = NULL,
    labels = NULL)
  
  se <- cl_predictive_comparison(
    se = se,
    subtypes = c("s", "r"),
    readouts = c("OS"),
    covariates = FALSE,
    fdr_i = 0.5)
  
  pl_example(se,
    mutations = "V4_SV",
    readout = "OS",
    subtype_column = "r",
    subtype = "r_cond_1",
    treatment = c("chemo", "no_chemo"))
}

}
