% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{pl_mutual_exclusivity}
\alias{pl_mutual_exclusivity}
\title{Plot mutual exclusivity}
\usage{
pl_mutual_exclusivity(se)
}
\arguments{
\item{se}{SE object}
}
\value{
ggplot object
}
\description{
Plot the results of the mutex algorithm
}
\examples{
\dontrun{
  sim_data <- sim_data()

  se <- prepare_data(
    data = sim_data,
    vars = c("OS", "DFS"),
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "PatientID",
    treatment_column = "Adj",
    mutation_columns = colnames(sim_data)[14:ncol(sim_data)]
  )

  pl_mutual_exclusivity(cl_mututal_exclusivity(se,
    min_variants = 10,
    mutex_output_exists = TRUE,
    save = "metadata/",
    mutex_path = "path/to/mutex/installation"
  ))
}

}
