% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{pl_oncoprint}
\alias{pl_oncoprint}
\title{Plot oncoprint}
\usage{
pl_oncoprint(se, min_mutants = 5, remove_mutations = NULL)
}
\arguments{
\item{se}{SE object}

\item{min_mutants}{mininum number of mutated tumours per mutation
to be included in the oncoprint}

\item{remove_mutations}{character vector of mutations to ignore for
oncoprint}
}
\value{
None
}
\description{
Plot oncoprint for interesting mutations
}
\examples{
{

  sim_data <- sim_data()

  se <- prepare_data(
    data = sim_data,
    vars = c("OS", "DFS"),
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "PatientID",
    treatment_column = "Adj",
    mutation_columns = colnames(sim_data)[14:ncol(sim_data)]
  )

  cl_subtype_enrichment(
    se = se,
    col_label = "Smoking.history",
    row_label = "EGFR.subtype",
    digits = 3
  )

  se <- cl_enrichment_genomics(se,
    sample_column = "sample",
    min_mutants = 10,
    subtype = c("Smoking.history", "EGFR.subtype")
  )
  
  pl_oncoprint(se)
}

}
