% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{pl_predictive_comparison}
\alias{pl_predictive_comparison}
\title{Plot treatment comparisons for predictive biomarkers}
\usage{
pl_predictive_comparison(
  se,
  min_altered = 10,
  max_fdr = 0.3,
  colors = NULL,
  labels = NULL
)
}
\arguments{
\item{se}{SE object}

\item{min_altered}{minimum number of mutant samples to be
tested for a treatment comparison}

\item{max_fdr}{FDR cutoff for subtype-specific tests}

\item{colors}{list of color specifications}

\item{labels}{list of label specifications}
}
\value{
list of ggplot objects
}
\description{
Plot treatment comparison tests in forest plots for the significant
interaction tests
}
\examples{
{

  sim_data <- sim_data()
  
  se <- prepare_data(
    data = sim_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(sim_data)[1:5])
    
    
  se <- cl_treatment_specific_biomarkers(se, 
     min_samples = 10,
     min_redistribution = 3,
     treatment = c("chemo", "no_chemo"),
     subtypes = c("s", "r"),
     readouts = c("OS"))
  
  se <- pl_treatment_specific_biomarkers_subtype(se, fdr_max = 0.1)
  
  se <- cl_tsb_modules_oncoprint(se, p_value = 0.05, fdr = 0.1)
  
  se <- cl_predictive_biomarkers(
    se = se,
    min_samples = 10,
    min_redistribution = 5)
  
  se <- pl_predictive_biomarkers_subtypes(se,
    fdr_max = 0.1,
    colors = NULL,
    labels = NULL)
  
  se <- cl_predictive_comparison(
    se = se,
    subtypes = c("s", "r"),
    readouts = c("OS"),
    covariates = FALSE,
    fdr_i = 0.5)
    
  pl_predictive_comparison(se)
}

}
