% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{pl_subtype_enrichment}
\alias{pl_subtype_enrichment}
\title{Plot subtype enrichment}
\usage{
pl_subtype_enrichment(se, limits = c(-4, 4))
}
\arguments{
\item{se}{SE object}

\item{limits}{plot color limits}
}
\value{
ggplot object
}
\description{
Plotting results after 'cl_subtype_enrichment'
}
\examples{
{
  
  simed_data <- sim_data()
  
  se <- prepare_data(
    data = simed_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(simed_data)[1:5])
    
  se <- cl_subtype_enrichment(
    se = se,
    col_label = "s",
    row_label = "r",
    digits = 3
  )
  
  pl_subtype_enrichment(se)
  
}

}
