% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{pl_treatment_specific_biomarkers}
\alias{pl_treatment_specific_biomarkers}
\title{Plot treatment-specific biomarkers}
\usage{
pl_treatment_specific_biomarkers(se)
}
\arguments{
\item{se}{SE object}
}
\value{
list of ggplot objects
}
\description{
Plots a volcano plot of treatment-specific biomarkers  for each
readout and treatment arm
}
\examples{
{

  sim_data <- sim_data()
  
  se <- prepare_data(
    data = sim_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(sim_data)[1:5]
    )
    
  se <- cl_treatment_specific_biomarkers(se, 
     min_samples = 10,
     min_redistribution = 3,
     treatment = c("chemo", "no_chemo"),
     subtypes = c("s", "r"),
     readouts = c("OS")
     )

  pl_treatment_specific_biomarkers(se)
}

}
