% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{pl_treatment_specific_biomarkers_subtype}
\alias{pl_treatment_specific_biomarkers_subtype}
\title{Plot treatment-specific and subtype-specific biomarkers}
\usage{
pl_treatment_specific_biomarkers_subtype(
  se,
  fdr_max = 0.3,
  colors = NULL,
  labels = NULL
)
}
\arguments{
\item{se}{SE object}

\item{fdr_max}{FDR cutoff}

\item{colors}{list of color specifications}

\item{labels}{list of label specifications}
}
\value{
SE object with metadata 'plot$treatment_specific_biomarkers_subtype'
as ggplot object
}
\description{
Plots a forest plot of treatment-specific and subtype-specific biomarkers
for each readout and treatment arm
}
\examples{
{

  sim_data <- sim_data()
  
  se <- prepare_data(
    data = sim_data,
    vars = "OS",
    med_impute_muts = TRUE,
    remove_clin = NA,
    sample_column = "ID",
    treatment_column = "trt",
    mutation_columns = colnames(sim_data)[1:5])
    
    
  se <- cl_treatment_specific_biomarkers(se, 
     min_samples = 10,
     min_redistribution = 3,
     treatment = c("chemo", "no_chemo"),
     subtypes = c("s", "r"),
     readouts = c("OS"))
  
  se <- pl_treatment_specific_biomarkers_subtype(se, fdr_max = 0.1)
}

}
