% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wrapper.R
\name{sim_data}
\alias{sim_data}
\title{Simulate Dataset}
\usage{
sim_data(n.obs = 400, n.vars = 5, n.subs = 3)
}
\arguments{
\item{n.obs}{number of observations}

\item{n.vars}{number of mutation variables}

\item{n.subs}{number of substitutions}
}
\value{
a dataframe which can be provided to PrepareData()
}
\description{
Simulate a dataset containig mutational and clinical data
}
\examples{
{

data <- sim_data()

data
}
}
